refact: extrair evasion/precision handlers para module/combat/evasion-precision.mjs
